+++ internal/cli/root.go
@@ -122,7 +122,7 @@ func init() {
 		}
 	})
 
-	rootCmd.AddCommand(serveCmd, daemonCmd, stopCmd, restartCmd, initCmd, setupCmd, skillsCmd, versionCmd, configCmd, sessionsCmd, mailCmd, statsCmd, diagnosticsCmd, doctorCmd, logLevelCmd, enableLSPCmd, debugCmd, webCmd)
+	rootCmd.AddCommand(serveCmd, daemonCmd, stopCmd, restartCmd, initCmd, setupCmd, skillsCmd, hooksCmd, versionCmd, configCmd, sessionsCmd, mailCmd, statsCmd, diagnosticsCmd, doctorCmd, logLevelCmd, enableLSPCmd, debugCmd, webCmd)
 	rootCmd.AddCommand(trustCmd)
 	rootCmd.AddCommand(taskCmds...)
 }